devex: move to ui/, remove graph and lib
also uninstall @urbit/api
