--- gear/storage/index.ts
@@ -1,2 +1,0 @@
-export * from './lib';
-export * from './types';
