+++ ui/gear/index.ts
@@ -1,7 +1,5 @@
 export * from './contacts';
 export * as contacts from './contacts';
-export * from './graph';
-export * as graph from './graph';
 export * from './groups';
 export * as groups from './groups';
 export * from './hark';
@@ -13,8 +11,8 @@ export * from './metadata';
 export * as metadata from './metadata';
 export * from './settings';
 export * as settings from './settings';
-export * from './s3';
-export * as s3 from './s3';
+export * from './storage';
+export * as storage from './storage';
 export * from './lib';
 export * from './lib/BigIntOrderedMap';
 export * from './lib/BigIntArrayOrderedMap';
